+++ src/libraries/System.IO.Compression.ZipFile/tests/ZipFile.Extract.cs
@@ -203,7 +203,7 @@ namespace System.IO.Compression.Tests
                     string filename = Path.Combine(tempFolder.Path, permission + ".txt");
                     Assert.True(File.Exists(filename));
 
-                    EnsureFilePermissions(filename, permission);
+                    ZipFileHelpers.EnsureFilePermissions(filename, permission);
                 }
             }
         }
